+++ bindgen-cli/options.rs
@@ -369,6 +369,9 @@ struct BindgenCommand {
     /// bitfields. This flag is ignored if the `--respect-cxx-access-specs` flag is used.
     #[arg(long, value_name = "VISIBILITY")]
     default_visibility: Option<FieldVisibilityKind>,
+    /// Whether to emit diagnostics or not.
+    #[arg(long)]
+    emit_diagnostics: bool,
     /// Enables experimental features.
     #[arg(long)]
     experimental: bool,
@@ -495,6 +498,7 @@ where
         wrap_static_fns_path,
         wrap_static_fns_suffix,
         default_visibility,
+        emit_diagnostics,
         experimental: _,
         version,
         clang_args,
@@ -997,12 +1001,25 @@ where
         }
     }
 
-    for (custom_derives, kind) in [
-        (with_derive_custom, None),
-        (with_derive_custom_struct, Some(TypeKind::Struct)),
-        (with_derive_custom_enum, Some(TypeKind::Enum)),
-        (with_derive_custom_union, Some(TypeKind::Union)),
+    for (custom_derives, kind, name) in [
+        (with_derive_custom, None, "--with-derive-custom"),
+        (
+            with_derive_custom_struct,
+            Some(TypeKind::Struct),
+            "--with-derive-custom-struct",
+        ),
+        (
+            with_derive_custom_enum,
+            Some(TypeKind::Enum),
+            "--with-derive-custom-enum",
+        ),
+        (
+            with_derive_custom_union,
+            Some(TypeKind::Union),
+            "--with-derive-custom-union",
+        ),
     ] {
+        let name = emit_diagnostics.then(|| name);
         for custom_derive in custom_derives {
             let (regex, derives) = custom_derive
                 .rsplit_once('=')
@@ -1011,7 +1028,7 @@ where
 
             let mut regex_set = RegexSet::new();
             regex_set.insert(regex);
-            regex_set.build(false);
+            regex_set.build_with_diagnostics(false, name);
 
             builder = builder.parse_callbacks(Box::new(CustomDeriveCallback {
                 derives,
@@ -1035,6 +1052,10 @@ where
 
     if let Some(visibility) = default_visibility {
         builder = builder.default_visibility(visibility);
+    }
+
+    if emit_diagnostics {
+        builder = builder.emit_diagnostics(true);
     }
 
     Ok((builder, output, verbose))